+++ geoopt/__init__.py
@@ -2,7 +2,7 @@ from . import manifolds
 from . import optim
 from . import tensor
 from . import samplers
-from . import util
+from . import linalg
 
 from .tensor import ManifoldParameter, ManifoldTensor
 from .manifolds import Stiefel, Euclidean, Sphere